refactor: 모집 공고 컴포넌트 글로벌화 및 모집 공고 스키마 추출 타입명 변경 (#96)
