+++ src/app/apply/_components/ApplyFormContainer.tsx
@@ -7,10 +7,10 @@ import {EtcInfo} from '@/app/apply/_components/EtcInfo';
 import {useApplyFormController} from '@/app/apply/_hooks/useApplyFormController';
 import {ApplicationConfirmModal} from '@/components/modal/ApplicationConfirmModal';
 import HeroMainBanner from '@/components/banner/HeroMainBanner';
-import {AdminRecruitmentInformation} from '@/app/admin/(with-sidebar)/application-edit/_components/recruitment/AdminRecruitmentInformation';
 import {useRecruitmentStatusQuery} from '@/hooks/queries/useRecruitmentStatus.query';
 import {useRecruitmentScheduleQuery} from '@/hooks/queries/useRecruitmentSchedule.query';
 import {Spinner} from '@/components/ui/Spinner';
+import {RecruitmentInformation} from '@/components/recruitment/RecruitmentInformation';
 
 const STEP_TITLES = {
   2: '파트별 질문',
@@ -62,7 +62,7 @@ export const ApplyFormContainer = () => {
               <span aria-hidden='true'>🥔</span>
             </h1>
             {scheduleData && (
-              <AdminRecruitmentInformation
+              <RecruitmentInformation
                 variant='plain'
                 data={scheduleData}
                 isEditing={false}
