+++ src/components/recruitment/RecruitmentInfoViewRow.tsx
